4.21*: bloglist expansion, step9

+++ part4/controllers/blogNotes.js
@@ -52,6 +52,18 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).json({ error: 'blog does not exist' })
+    }
+
+    if (!request.token || !decodedToken.id || !user || (blog.user.toString() !== user.id)) {
+        return response.status(401).json({ error: 'no permission to delete this blog' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
